+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Add feature to replace equipment with elite aspects
+- Fix link in readme to the original mod for real
+
 # 1.0.1
 
 - Fix link in readme to the original mod